Refactor with pages/ directory (#7)
