+++ src/components/index.js
@@ -1,10 +1,7 @@
 export * from './AuthorSelect';
-export * from './ErrorPage';
-export * from './HomePage';
 export * from './IntersectionsLayer';
 export * from './ItinerariesLayer';
 export * from './ITSBStore';
 export * from './ITSBTooltip';
-export * from './MapPage';
 export * from './MonthRangeInput';
 export * from './Header';
