+++ src/pages/index.js
@@ -1,0 +1,3 @@
+export * from './ErrorPage';
+export * from './HomePage';
+export * from './MapPage';
